# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GPUBROKER API (Current Endpoints)
  version: 0.2.0
  description: >
    Canonical minimal contract for Auth, Provider, KPI, Booking preview, and WebSocket price updates.
    Keep in sync with FastAPI /docs outputs; expand as services grow.
servers:
  - url: http://localhost:28020
  - url: http://localhost:28021
  - url: http://localhost:28022
  - url: http://localhost:28026
  - url: ws://localhost:28025
paths:
  /auth/register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Registration accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  verification_required:
                    type: boolean
        '400':
          description: Invalid input
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
        '401':
          description: Unauthorized
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT